Use bash in subprocess calls in Linux/Unix
When using GNU parallel to run third-party software in parallel (i.e. multi-processing) on Linux the default behavior of subprocess is to use /bin/sh. This clashes with the syntax used for the model_command_string as /bin/sh shell cannot correctly unpack the numerical sequence {0..n} at the end of the GNU parallel command. Therefore, it will try to change directory into a non-existing directory (for example "cd run_{0..10}", where here n=10). This will produce an error as the GNU parallel command will not be correctly executed and the model will not be run in parallel. By simply checking if the system platform is Linux and explicitly telling subprocess to use /bin/bash instead of /bin/sh this error is fixed